gpupdate-setup: add support domain_controller local policy profile
